modify layouting and navbar

+++ client/src/components/NavbarHome.vue
@@ -1,22 +1,36 @@
 <script>
 import {mapWritableState} from "pinia"
 import {useCounterStore} from "../stores/counter"
+import Swal from 'sweetalert2'
 export default {
   created(){
     if(localStorage.access_token){
       this.isLoggedIn = true
-      this.subscribe = localStorage.isSubscribe
     }
+    this.subscribe = localStorage.isSubscribed
   }, 
   computed: {
     ...mapWritableState(useCounterStore, ["isLoggedIn", "subscribe"])
   },
   methods: {
     logoutHandler(){
-      localStorage.clear()
-      this.$router.push("/login")
+      Swal.fire({
+        title: "Are you sure?",
+        text: "You want to log out from Newkey?",
+        icon: "warning",
+        showCancelButton: true,
+        confirmButtonColor: "#3085d6",
+        cancelButtonColor: "#d33",
+        confirmButtonText: "Yes, log out!",
+      }).then((result) => {
+        if (result.isConfirmed) {
+          localStorage.clear();
+          this.isLoggedIn = false;
+          this.$router.push("/")
+        }
+      });
     }
-  }
+  },
 }
 </script>
 
@@ -24,13 +38,12 @@ export default {
   <div id="box3" v-if="isLoggedIn">
     <div id="container">
       <div id="logo">
-        <img src="logo" />
       </div>
       <div id="menu">
         <ul>
           <router-link to="/lobby" class="router">LOBBY</router-link>
-          <router-link v-if="!subscribe" to="/subscription" class="router">SUBSCRIBE</router-link>
-          <a to="/logout" @click.prevent="logoutHandler" class="router">LOGOUT</a>
+          <router-link v-if="subscribe == 'false'" to="/subscription" class="router">SUBSCRIBE</router-link>
+          <a @click.prevent="logoutHandler" class="router">LOGOUT</a>
         </ul>
       </div>
     </div>
@@ -77,4 +90,6 @@ export default {
   display: inline-block;
   margin-left: 100px;
 }
+
+
 </style>
